WebBoxFashion/requirements.txt
aniso8601==8.0.0
astroid==2.3.1
attrs==19.3.0
Click==7.0
Django==2.2.6
django-solo==1.1.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.7
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
freeze==1.0.10
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonify==0.5
jsonschema==3.1.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
Pillow==6.2.0
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pylint==2.4.2
pyrsistent==0.15.4
python-dotenv==0.10.3
pytz==2019.3
six==1.12.0
SQLAlchemy==1.3.10
sqlparse==0.3.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
